Extracts Drop model to models file
