+++ src/reducers/drop.tsx
@@ -1,10 +1,10 @@
 import * as types from "../actions/actionTypes";
 import { Locations, Houses } from "../constants";
-import { Piece } from "../models";
+import { Drop } from "../models";
 
 interface DropState {
   isLoading: boolean;
-  drops: Piece[];
+  drops: Drop[];
 }
 
 const initialState: DropState = {
@@ -14,12 +14,12 @@ const initialState: DropState = {
 
 export default (
   state = initialState,
-  { type, drop }: {type: string, drop: Piece}
+  { type, drop }: {type: string, drop: Drop}
 ) => {
   switch (type) {
     case types.RESET_COMBAT_SUCCESS:
-      const dropsInCombat: Piece[] = state.drops.filter((p: Piece) => p.location === Locations.COMBAT)
-      const resetedDrops: Piece[] = dropsInCombat.map(d => ({
+      const dropsInCombat: Drop[] = state.drops.filter((p: Drop) => p.location === Locations.COMBAT)
+      const resetedDrops: Drop[] = dropsInCombat.map(d => ({
         ...d,
         location: Locations.WAR_ROOM,
         spec: {
@@ -31,7 +31,7 @@ export default (
       return {
         isLoading: false,
         drops: [
-          ...state.drops.filter((p: Piece) => p.location !== Locations.COMBAT),
+          ...state.drops.filter((p: Drop) => p.location !== Locations.COMBAT),
           ...resetedDrops,
         ],
       };
@@ -44,7 +44,7 @@ export default (
       return {
         isLoading: false,
         drops: [
-          ...state.drops.filter((p: Piece) => p.id !== drop.id),
+          ...state.drops.filter((p: Drop) => p.id !== drop.id),
           {
             ...drop,
             spec: flippedSpec,
@@ -61,7 +61,7 @@ export default (
       return {
         isLoading: false,
         drops: [
-          ...state.drops.filter((p: Piece) => p.id !== drop.id),
+          ...state.drops.filter((p: Drop) => p.id !== drop.id),
           {
             id: drop.id,
             type: drop.type,
